+++ Idealworld/src/main/java/com/yjy/idw/Controller/ResultController.java
@@ -1,0 +1,12 @@
+package com.yjy.idw.controller;
+/**
+ * url : localhost:8080/results
+ */
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/results")
+public class ResultController {
+
+}
