+++ Idealworld/src/main/java/com/yjy/idw/Controller/TournamentController.java
@@ -1,0 +1,12 @@
+package com.yjy.idw.controller;
+/**
+ * url : localhost:8080/tournaments
+ */
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/tournaments")
+public class TournamentController {
+
+}
