+++ Idealworld/src/main/java/com/yjy/idw/Controller/UserController.java
@@ -10,9 +10,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.yjy.idw.UserService;
-import com.yjy.idw.UserVO;
-
+import com.yjy.idw.user.UserService;
+import com.yjy.idw.user.UserVO;
+/**
+ * url : localhost:8080/users/{id}
+ * @author yunbok
+ *
+ */
 @RestController
 @RequestMapping("/users")
 public class UserController {
@@ -34,7 +38,7 @@ public class UserController {
 		userService.deleteUser(Integer.parseInt(id));
 	}
 	
-	@RequestMapping(method = RequestMethod.POST, value = "") 
+	@RequestMapping(method = RequestMethod.POST) 
 	public String insertUser(UserVO vo) {
 	   	userService.insertUser(vo);
 		return "추가완료";
