+++ Idealworld/src/main/java/com/yjy/idw/result/ResultDAO.java
@@ -1,0 +1,5 @@
+package com.yjy.idw.result;
+
+public class ResultDAO {
+
+}
