+++ Idealworld/src/main/java/com/yjy/idw/result/ResultService.java
@@ -1,0 +1,5 @@
+package com.yjy.idw.result;
+
+public interface ResultService {
+
+}
