+++ Idealworld/src/main/java/com/yjy/idw/tournament/TournamentDAO.java
@@ -1,0 +1,5 @@
+package com.yjy.idw.tournament;
+
+public class TournamentDAO {
+
+}
