+++ Idealworld/src/main/java/com/yjy/idw/tournament/TournamentVO.java
@@ -1,0 +1,5 @@
+package com.yjy.idw.tournament;
+
+public class TournamentVO {
+
+}
