+++ Idealworld/src/main/java/com/yjy/idw/user/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.yjy.idw;
+package com.yjy.idw.user;
 
 import java.util.List;
